+++ sdk-api-kotlin-gen/src/main/kotlin/dev/restate/sdk/kotlin/gen/ComponentProcessorProvider.kt
@@ -1,0 +1,21 @@
+// Copyright (c) 2023 - Restate Software, Inc., Restate GmbH
+//
+// This file is part of the Restate Java SDK,
+// which is released under the MIT license.
+//
+// You can find a copy of the license in file LICENSE in the root
+// directory of this repository or package, or at
+// https://github.com/restatedev/sdk-java/blob/main/LICENSE
+package dev.restate.sdk.kotlin.gen
+
+import com.google.devtools.ksp.processing.SymbolProcessor
+import com.google.devtools.ksp.processing.SymbolProcessorEnvironment
+import com.google.devtools.ksp.processing.SymbolProcessorProvider
+
+class ComponentProcessorProvider : SymbolProcessorProvider {
+
+  override fun create(environment: SymbolProcessorEnvironment): SymbolProcessor {
+    return ComponentProcessor(
+        logger = environment.logger, codeGenerator = environment.codeGenerator)
+  }
+}
