+++ sdk-api-kotlin-gen/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,1 +1,1 @@
-ComponentProcessorProvider
+dev.restate.sdk.kotlin.gen.ComponentProcessorProvider
